refactor(migrations): Move ChangeTracker to common utils
The `ChangeTracker` is generally useful and could be used by a lot of
migrations instead of having to rewrite similar boilerplate.
